Added CellGrid supertype to Tile and MultiBandTile.
