+++ raster/src/main/scala/geotrellis/raster/CellGrid.scala
@@ -1,0 +1,11 @@
+package geotrellis.raster
+
+trait CellGrid {
+  def cols: Int
+  def rows: Int
+  def cellType: CellType
+
+  def dimensions: (Int, Int) = (cols, rows)
+  def gridBounds: GridBounds = GridBounds(0, 0, cols - 1, rows - 1)
+  def size = cols * rows
+}
