+++ spark/src/main/scala/geotrellis/spark/RasterMetaData.scala
@@ -25,11 +25,11 @@ case class RasterMetaData(
 }
 
 object RasterMetaData {
-  def envelopeExtent[T](rdd: RDD[(T, Tile)])(getExtent: T => Extent): (Extent, CellType, CellSize) = {
+  def envelopeExtent[K, V <: CellGrid](rdd: RDD[(K, V)])(getExtent: K => Extent): (Extent, CellType, CellSize) = {
     rdd
-      .map { case (key, tile) =>
+      .map { case (key, grid) =>
         val extent = getExtent(key)
-        (extent, tile.cellType, CellSize(extent, tile.cols, tile.rows))
+        (extent, grid.cellType, CellSize(extent, grid.cols, grid.rows))
       }
       .reduce { (t1, t2) =>
         val (e1, ct1, cs1) = t1
@@ -45,8 +45,8 @@ object RasterMetaData {
   /**
    * Compose Extents from given raster tiles and fit it on given [[TileLayout]]
    */
-  def fromRdd[T](rdd: RDD[(T, Tile)], crs: CRS, tileLayout: TileLayout)
-                (getExtent: T => Extent): RasterMetaData = {
+  def fromRdd[K, V <: CellGrid](rdd: RDD[(K, V)], crs: CRS, tileLayout: TileLayout)
+                (getExtent: K => Extent): RasterMetaData = {
     val (uncappedExtent, cellType, cellSize): (Extent, CellType, CellSize) = envelopeExtent(rdd)(getExtent)
     val worldExtent = crs.worldExtent
     val extentIntersection = worldExtent.intersection(uncappedExtent).get
@@ -57,8 +57,8 @@ object RasterMetaData {
    * Compose Extents from given raster tiles and pick the closest [[LayoutLevel]] in the [[LayoutScheme]].
    * @param isUniform   all the tiles in the RDD are known to have the same extent
    */
-  def fromRdd[T](rdd: RDD[(T, Tile)], crs: CRS, layoutScheme: LayoutScheme, isUniform: Boolean = false)
-                (getExtent: T => Extent): (LayoutLevel, RasterMetaData) = {
+  def fromRdd[K, V <: CellGrid](rdd: RDD[(K, V)], crs: CRS, layoutScheme: LayoutScheme, isUniform: Boolean = false)
+                (getExtent: K => Extent): (LayoutLevel, RasterMetaData) = {
     val (uncappedExtent, cellType, cellSize): (Extent, CellType, CellSize) =
       if(isUniform) {
         val (key, tile) = rdd.first
@@ -71,6 +71,6 @@ object RasterMetaData {
     val worldExtent = crs.worldExtent
     val layoutLevel: LayoutLevel = layoutScheme.levelFor(worldExtent, cellSize)
     val extentIntersection = worldExtent.intersection(uncappedExtent).get
-    layoutLevel -> RasterMetaData(cellType, extentIntersection, crs, layoutLevel.tileLayout)
+    (layoutLevel, RasterMetaData(cellType, extentIntersection, crs, layoutLevel.tileLayout))
   }
 }
